requirements.txt
asciimatics==1.12.0
attrs==20.3.0
cached-property==1.5.2
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
commonmark==0.9.1
contextvars==2.4
crayons==0.4.0
cryptography==3.2.1
culour==0.1
Cython==0.29.21
dataclasses==0.7
decorator==4.4.2
dill==0.3.3
dlx==1.0.4
docplex==2.18.200
fastdtw==0.3.4
fastjsonschema==2.14.5
func-timeout==4.3.5
future==0.18.2
h5py==3.1.0
idna==2.10
immutables==0.14
importlib-metadata==3.1.0
inflection==0.5.1
joblib==0.17.0
jsonschema==3.2.0
lxml==4.6.2
more-itertools==8.6.0
mpmath==1.1.0
multitasking==0.0.9
nest-asyncio==1.4.3
networkx==2.5
ntlm-auth==1.5.0
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
pkg-resources==0.0.0
playsound==1.2.2
ply==3.11
psutil==5.7.3
py-cui==0.1.2
pybind11==2.6.1
pycparser==2.20
pyfiglet==0.8.post1
Pygments==2.7.2
pyrsistent==0.17.3
python-constraint==1.4.0
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
qiskit==0.23.1
qiskit-aer==0.7.1
qiskit-aqua==0.8.1
qiskit-ibmq-provider==0.11.1
qiskit-ignis==0.5.1
qiskit-terra==0.16.1
qrng==0.1.2
Quandl==3.5.3
requests==2.25.0
requests-ntlm==1.1.0
retworkx==0.6.0
rich==9.1.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sympy==1.7
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
websockets==8.1
yfinance==0.1.55
zipp==3.4.0
